- Scenario support completed. - Some bugfixes for the scheduler. - Export files have now correct dependencies. - CoreAttributes::getType returns now a number and not a string.
git-svn-id: https://www.taskjuggler.org/svn/taskjuggler/trunk@523 e1914e07-63f8-0310-9059-d6d858d7cdca

+++ taskjuggler/TaskList.cpp
@@ -53,7 +53,7 @@ TaskList::compareItemsLevel(Task* t1, Task* t2, int level)
     if (level < 0 || level >= maxSortingLevel)
         return -1;
 
-    int sc = sorting[level] >> 16;
+    int sc = ((unsigned int) sorting[level]) >> 16;
     switch (sorting[level] & 0xFFFF)
     {
     case TreeMode:
